requirements.txt
openai-whisper==20250625
openai==1.97.1
python-dotenv==1.0.1